tests: colocate P1 fixture coverage with owners
